Added DAO methods to insert/update words in batch and modified importers to use batch methods.  Refactored Word entity to use int instead of long for better memory efficiency.  Minor data cleanup.

+++ src/main/java/com/ciphertool/sentencebuilder/entities/Word.java
@@ -9,7 +9,7 @@ import javax.persistence.Table;
 @Table(name="parts_of_speech")
 public class Word {
 	private WordId wordId;
-	private long frequencyWeight;
+	private int frequencyWeight;
 	
 	public Word() {}
 	
@@ -33,11 +33,11 @@ public class Word {
 	}
 	
 	@Column(name="frequency_weight")
-	public long getFrequencyWeight() {
+	public int getFrequencyWeight() {
 		return frequencyWeight;
 	}
 	
-	public void setFrequencyWeight(long frequencyWeight) {
+	public void setFrequencyWeight(int frequencyWeight) {
 		this.frequencyWeight = frequencyWeight;
 	}
